Extract `:encryption-services` out of `:configuration-cache`
