+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/internal/cc/impl/ConfigurationCacheKey.kt
@@ -19,6 +19,7 @@ package org.gradle.internal.cc.impl
 import org.gradle.internal.extensions.stdlib.unsafeLazy
 import org.gradle.internal.cc.impl.initialization.ConfigurationCacheStartParameter
 import org.gradle.internal.buildtree.BuildActionModelRequirements
+import org.gradle.internal.encryption.EncryptionConfiguration
 import org.gradle.internal.hash.Hasher
 import org.gradle.internal.hash.Hashing
 import org.gradle.internal.service.scopes.Scope
