+++ platforms/core-configuration/core-kotlin-extensions/build.gradle.kts
@@ -11,6 +11,7 @@ dependencies {
     api(projects.coreApi)
     api(projects.files)
     api(projects.hashing)
+    api(projects.loggingApi)
 
     api(libs.kotlinStdlib)
