+++ platforms/core-configuration/encryption-services/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,0 +1,1 @@
+org.gradle.internal.encryption.services.EncryptionServices
